requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.22.0
Django==3.1.4
django-bootstrap4==2.3.1
django-cloudinary-storage==0.3.0
django-crispy-forms==1.9.2
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
idna==2.10
Markdown==3.3.3
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.4
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
